Graph Model: Move graph context implementation into base class
Instead of reimplementing graph context for every tool and test, upgraded the graph context interface into a base class and moved the implementation into it.
Preserve the virtual functions necessary for landscape canvas special entity ID logic.
Added functions for looking up types by name.
Changed function that looked up types by enum to search instead of direct index because nothing is enforcing that types are added in the same order as the enumeration. This might be more optimal with an associative container type.
Added construction parameters to pass in the system name, module details, and data types so the class does not need to be overridden in most cases.
Still need to address the case where the module manager relies on the graph context which means it cannot be constructed in the constructor.

Signed-off-by: Guthrie Adams <guthadam@amazon.com>